requirements.txt
diffusers
transformers
streamlit
Pillow